Remove set heights from html/body and overlay
If the webpage is too big to fit into the window, avgrund messes up the
overlay and virtually disables scrolling on chrome with jquery or
javascript for that matter. For this purpose the 100% height on html and
body was removed and overlay now retreives the actual height of the
body, not some insane 20000px value.